requirements.txt
streamlit==1.52.2
Pyrebase4==4.9.0
spacy==3.8.11
requests==2.32.5
streamlit-lottie==0.0.5
PyPDF2==3.0.1
python-docx==1.2.0
scikit-learn==1.8.0
groq==0.15.0
